Add phase adjust / deadband for PID controller
Currently the PID loop always uses either + or - maximum phase advance.
This means that when the PID output value is 0, there is no torque applied.
This causes the motor to zoom past the setpoint if there is no external friction
to stop it, which in turn causes a lot of oscillations.

This commit adjusts the phase angle when PID output is close to zero, which
acts to brake the motor and allows it to stop accurately between steps also.